implemented Disc Caching for the json response

+++ src/main/java/com/example/MyWatchList/DataModels/MovieModels/MovieInfoPageModel.java
@@ -6,14 +6,14 @@ public class MovieInfoPageModel extends MediaInfoPageModel {
     private BelongsToCollection belongs_to_collection;
     private int budget;
     private String release_date;
-    private int revenue;
+    private long revenue;
     private int runtime;
     private String title;
 
     public BelongsToCollection getBelongs_to_collection() {return belongs_to_collection;}
     public int getBudget() {return budget;}
     public String getRelease_date() {return release_date;}
-    public int getRevenue() {return revenue;}
+    public long getRevenue() {return revenue;}
     public int getRuntime() {return runtime;}
     public String getTitle() {return title;}
